Fix next/previous buttons issue in Media Player app example
If we loaded several videos, clicking on the next button
moved on to the next video, but if we did this on the last video,
the app would select an 'empty' video and only on the second click
would it move to the front of the list.
The same was happening with the previous button.

The problem was the function that determines
the index of the next preceding video, so I fixed it.

Fixes: QTBUG-119083
Change-Id: I6fc573cfc09338c8c8dc4aef7f8d0b8a6e74290e
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
(cherry picked from commit 209eec6894479a3aa83c778c488ea4986d5c38ec)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d470f0545046e3ced90c3a652961f9b5bdf6b0cc)